Address review feedback
- move mono specific stuff to `MSBuildLocator.Mono.cs`
